UDT connect issue fixed
Wow. In CPacket I changed uin32_t to int32_t for base packet header type.
(4x 32bit numbers) because those are accessed through int32_t references
most of the time anyway.

The method to extract a flag from one of the header fields (bit 31) used
this construction:

    int CPacket::getFlag() {
        return m_nHeader[0] >> 31;
    }

What happened was that because m_nHeader[] now was int32_t, the ">>"
operator will sign-extend to the right, so with:
    m_nHeader[0] == 0x80000000
">> 31" will result in  this:
    0xffffffff
and be returned as int, i.e. as "-1", in stead of the *intended* "1".

The main code checks - verbatim code here - like this:

  if ((1 != response.getFlag()) || ... )

which fails miserably because on int comparison "1 != -1" ....