Officially support remote vbs->vbs wildcards
The built-in scan chunk finder in jive5ab was flexible enough to deal with
patterns; it was mostly m5copy preventing them to be used on a remote
flexbuff.

Introduced a separate data source for vbs->vbs transfers in m5copy that does
not barf on wildcards but instead transforms shell-stye globbing to regex
before passing that pattern on to the jive5ab's it is talking to.

The regex inside jive5ab was made a little better - it now literally only
matches official scan chunks, even if a wildcard was given. (In the previous
version it could match more than actually intented).