handle suid root jive5ab, add --allow-root
Normally, jive5ab drops all its privileges, even if it is run suid root. So,
normally, all files created by jive5ab are still owned by the caller of
jive5ab.

But sometimes jive5ab must be able to read from privileged ports - the RDBEs
send UDP packets to port 625, which is unconfigurable. Making jive5ab suid
root makes that fixable - combined with NOT dropping privileges, easy
enough.

The downside is that all files then created are owned by root, which is not
very handy to work with.

This fix applies an ownership change to the caller's ownership in such a
case. This only happens if the effective user id of jive5ab is root; only
root can change ownership of a file.