scan_set= with time can fail using only ~1MB
When analyzing VDIF data at high speeds, the default, compiled in ~1MB of
data scan_set= analyzes to detect the data format (in order to be able to
support time-based data selection) can be too small.

The culprit is "is_data_format()" looking for VDIF. It requires a next frame
of the *same* thread to be found in the block of data to analyze. At high
data rates (e.g. VGOS) chances are that within that 1MB no frame of the same
thread as the first frame will be found.

In such a case (no frame of same thread found), it doesn't detect a data
format at all, making the whole command fail.