module init functionality, other improvements
m6sg_mount can now be used to (re)initialize (a) Mark6 disk pack(s)

    m6sg_mount --init [1234]+ --msn 1=MSN1 ... --format {xfs|ext4}

It is possible to initialize multiple disk packs in one go; it is
verified that for each module a MSN is specified; overspecification
(setting an MSN for a module not being initialized) is just a warning.

Each disk will be repartitioned into two partitions:
    1. data (whole disk less ~100MB)
    2. meta data (the remaining ~100MB)

The partitions can be formatted with xfs or ext4 file systems.

The eMSN file is generated and written to the meta-data partitions.
On the data partitions the "data/" directory is created.

Other improvements are:
    - the mount points for the data, meta-data partitions are
      created if not already exist
    - code clean up
    - improved subprocess handling
    - some formatting improvements
    - fixed default handling when user requested to type
      an answer -> that code didn't work at all!
      Now the default actually works as intended (i.e. do NOT
      proceed with action!)