Fix missing partial last block + larger buffers
Q[SimonC+EskilV]: Oi, when recording FlexBuff, we don't see any partial last
blocks when "record=off" is given. Where's the data that's been collected so
far gone?

A[me]: Wut? You must be doing it wrong because we use "gentle_stop()" on the
recording to make sure all pushed data flows downstream and also the code is
*allowed* to handle partial block (there is a setting for that).

[SimonC+EskilV]: You, sir, are an idiot because it really doesn't work!

[me, really thinking about what happens]: But of course you are right! I AM
an idiot!

    * network reader fills blocks and only (attempts) to push:
        - when block is full
        - when socket is closed, try to push whatever was buffered so far,
          if anything

    * record=off does ".gentle_stop()". The purpose of that is to
      first disable pushing on the producer's queue such that any steps
      following it get a chance to process the data in all queues,
      before stopping.

See there's the problem: the network reader breaks from its loop
because the queue it pushes on is disabled, meaning that any partial
blocks it holds are lost.

Fix: special handling of network readers in net2vbs. When record=off is
given, *first* call close_filedescriptor(), then wait 0.5s before disabling
the queues (gentle_stop()).

So now the network reader breaks from the loop because the network's gone
away and it can still push the partial block(s) downstream before stopping.

Whilst fixing that, made sure that the number of queue positions is at least
as large as either the number of mountpoints or the number of disk writer
threads, whichever is larger. This should increase the chance that all data
can be cached to memory and flushed to disk subsequently.

Will fail in case the incoming data rate is >> larger than sustained disk
write speed such that all queue positions are always full. Then again, that
is something that cannot be avoided.