Fix assert on wrong variable!
Guifre's compiler ("V3.0.0-rc1-hotfix failed in Debian 10.3 – Compiler version 8.3.0")
IS ONE OF THE SHARPER TOOLS IN THE SHED!! Good catch this one:

src/mk6info.cc:247:53: error:
‘*<unknown>.vdif_key::<anonymous>.vdif_key::<unnamed union>::station_code[0]’
may be used uninitialized in this function:
     EZASSERT2( ::strlen(code) <= 2 && station_code[0]!='\0', datastreamexception_type,
                                                     ^

Which is just a warning (and because of our very strict compilation
flags thus an error) but the compiler is absolutely right!

The assert should check the constructor argument, not the as-yet
uninitialized data member!