Some more throw() specifications updated
The throw() specifications are very different for C++98 and C++11+
Compilers increasingly struggle with the old syntax (which we use), more
specifically, the default compilers for modern systems assume C++11
which actually has a different /behaviour/ of how to e.g. deal with throwing
destructors.

At some point we may need to go to 11, because the old-style throw()
specifications will be removed from the language.

But for now let's hope these mods keep the compilers happy for a bit
longer.