Document the risks/antimeasures of m5copy-X.YY
The single-file checkout is not entirely w/o risks. Better document that.
We're definitely open to suggestions of how to do this better.

The underlying 'problem' is that git tags (and versions) apply to the whole
code base and not to individual files. In > 99% of the use cases this is a
desirable property, but not always.