Remove empty file
If, at the end of a net2file transfer, the destination file is empty, we
remove it - NO QUESTIONS ASKED.

Reason: the code currently does not know if it actually created the
file, or if it was pre-existing and we were asked to a(ppend) or w(rite)
- i.e. OVERwrite (truncate to zero size).

Net result is that if a file of size zero pre-existed and a(ppend) was
requested and no data transferred: it will be zapped.

If a file of non-zero size pre-existed and w(rite) was requested, this
truncates the file to zero first and if no bytes were transferred, the
file is zapped.

This could lead to premeditated removal of remote files by using a data
connection that is known to fail.