deadlock 101!?!
This commit:
https://github.com/jive-vlbi/jive5ab/commit/0b98e0dfbddc949a1992d05e8262dc3503855730

introduced a patch in `chain::run_step()` that causes
`chain::do_finals()` to be executed with the chain's mutex held.

Most of the final functions ("guard functions" in jive5ab) call
"communicate(...)" on the chain, and the first thing that that one does
is ... grab the mutex. Ergo deadlock.

After analysis with BE seems that that whole patch is maybe wonky, but
in order to prevent deadlocking undoing this particular one already goes
a long way.