Enable control of unique name check (issue #17)
At "record=on:<scan name>" on FlexBuff/Mark6 jive5ab would
unconditionally check for a pre-existing recording of the same "<scan
name>" and find and append a unique suffix to the scan name to record
in before actually starting capturing data from the network.
Recording data into a pre-existing scan will compromise the recording
and cause loss of both old and new recording.

However ... on a filled FlexBuff/Mark6 this can take relatively long time (~0.5s)
which, on 3s scans, is a long time.

This commit enables full control over whether the check is performed or
not. If the controlling system can guarantee unique scan names (e.g.
FieldSystem) the check for unique names can now be switched off:

- on the jive5ab command line the global default for all runtimes can be
  changed; the compiled-in default is "true", i.e. DO check for
  pre-existing recordings and change the scan name with a suffix

- in each runtime the setting can be dynamically controlled:
    record = check_unique_recording_names : [01] ;
  to explicitly turn the check off (0) or on (1)

- Use "record ? check_unique_recording_names ;" to enquire
  about the current setting