Fix broken assert condition
After v3.0.0 tagging "scanChunkFinder" implementation was changed and
the error condition to verify if a mount point could be opened altered.

This code was not tested in production (obviously) because the condition
was implemented completely wrong: it required opendir() to return a NULL
pointer, which, under normal circumstances, would never do because it
should only attempt to read mountpoints it has access to.

Net result: if everything was working successfully, the assert would
trigger, making the code not find any scan chunks, ever.