Fix scan_set with data stream labels
The idea is that with multi-file recordings, "scan_set=<scan base name>"
will *also* find the multi files produces with "_dsXXXX" data stream
label.
Most of that instrumentation was implemented but didn't work because the
data stream label turned out to not be extracted from the file name
correctly. Now it's a bit better - at least it finds the multiple
recordings of a scan.

(the first part of the match turned out to be too gready; it et the data
stream aux info field too).