new scan_check? algorithm also enable quiet run
The scan_check? algorithm was really simple: read a bit of data
(default: 1MB) at begin and end and that was enough.
With (very) high data rate multi-thread VDIF this approach broke down
since it's unlikely to see all VDIF threads nor a frame number wrap.

Algorithm reworked to, in case of VDIF, sample the recording at multiple
points and gather all information before trying to work out actual data
rate or number of VDIF threads.

The scan_check was very "noisy" on the jive5ab terminal if message level
>= 4 and compiled in Debug mode so added a possibility to shut it up.
This was mostly errors from time stamp decoding - the code just trying a
variety of data formats; this info only useful if debugging the
*timestamp* decoding.

"scan_check" now has a settable parameter: verbose operation (default:
Yes to be backwards compatible.

Use:
    scan_check = verbose : [01];

to set it No or Yes.

    scan_check ? verbose ;

can be used to check the current state of verbosity.