scan_check: fix missing bytes probs
Turns out there were two problems remaining with missing bytes
computation:

- read_offsets of the sampling points (for VDIF) were integer multiples
of the "read size" (default ~1MiB); the number of bytes to examine at
each sampling point. This meant that on average ~1/2 of such a block
would not be accounted for, leading to very weird missing bytes (in
fact: extra bytes) being reported. This is now fixed; the last sampling
point is always "read size" before the end of the file, rounded to the
nearest maximum integer number of frames that will fit in such a block.

- after that was fixed a small discrepancy remained between
scan_checking a multiple datastream recording in one go or in parts: in
one case one frame too much found (the former, combined recording)
whilst the constituent recordings (the latter, each thread scan_check'ed
individually) reports no bytes lost at all. Fixed by realising that the
byte offset and time stamp reported for the last frame in the recording
correspond to the _start_ of that frame. By including an extra frame in
duration of the scan (the samples in the file actually cover _that_
duration) and one frame's worth of bytes (the file extends one
framelength beyond the offset of the last frame header) both cases now
scan_check perfectly with zero missing bytes.

Note: the assumption is that that last frame is written completely in
the file; not an unreasonable assumption, and if that condition is'nt
true, the code will actually report missing bytes, which in this case
it's fully entitled to: the frame wasn't written to file completely ...