Add more detail if scan_check fail
When looking at VDIF the code samples at several points in the recording
trying to aggregate as much info as possible.
If the code did not find the same "format" at some point as at the start
(same "shape" of VDIF - for some value of "shape") it #FAILs (which is
OK), but the error message was very uninformative WHY.

This commit effectuates that more VDIF details are carried to a higher
layer (the scan_check implementation) such that it can describe why what
it found was not what it expected.

BeppeM produced a multi-stream VDIF recording where one thread was 2
channels x 8 bits/sample and the other was 8 channels x 2 bits per
sample.

Number of "bit streams" are same but (logical) content is different.
That (logical) info is now carried through and included in the error
message. (It took me quite some debugging time to figure out *why* the
scan check #FAILed ...)