Forgot closing } in C++11 compilation
The code was only compiled when linking with e-transfer, #ifdef'ed on
C++11 and I had forgotten closing curly braces. Took some time to find
where ...

Similarly, the move constructor was incorrectly coded (again only when
compiling w/c++11), and because of the move c'tor, required a copy c'tor
as well (it was implicitely deleted b/c of the manually implemented move
c'tor).