Extend net_port= to multiple ports and suffixes
The "net_port = [<host|ip>@]port" command is used to set the network
communication port for incoming/outgoing network connections, and
has supported specifying a local address to bind(2) to for listening
on for incoming connections/data.

This is a per-runtime specific setting and was limited to one (1) port.

This patch extends that paradigm to this syntax:

    net_port = [addr1@]port1[=suffix1] [ : [addr2@]port2[=suffix2] : ... ] ;

Basically allowing for:
    - specifying multiple ports for incoming data
    - specifying an optional suffix to append to
      the recording name set by "record = on : <name> ;"

Note: all local addresses ("addr1", "addr2", ...), if not specified,
default to IPv4 "any local address" (0.0.0.0 in dotted-quad notation),
i.e. all IPv4 interfaces on the current host.

Example: if jive5ab is configured like this:

    net_port = [addr1@]port1=suffix1 : [addr2@]port2=suffix2 ;

then, after "record = on : <name> ;", data read from:

    - [addr1]port1 is written to a recording "<name>_ds<suffix1>"
    - [addr2]port2 is written to a recording "<name>_ds<suffix2>"

Note: this patch only prepares the internal bookkeeping, the actual
commit that updates all reader code to make this happen follows.