transparent multistream reading in one fell swoop!
This is "the big push". A combination of code reorganisation and the
implementation of supporting a long-planned feature:

    net_port = [<addr1>@]port1[=<suffix1>] : [<addr2>@]port2[=<suffix2>] ;
    record = on : <name> ;

to transparently read from multiple network ports at the same time and
separate/tag the traffic by port. For each defined net_port a separate
thread is started to capture all incoming data.  This should make
breaking the ~16-20 Gbps single-core network capturing limit trivially,
by very easily engaging multiple cores.
Note that the sending digital backends need to be reprogrammed to send
data to different IP addresses and/or port numbers.

This commit encompasses several code changes:

-- making functions multiple port aware
Because of the change of runtime->host/port 1:1 going to 1:n all
functions dealing with a runtime's host and/or port setting changed:

    - now use host & port index #0 (there will always be at least one
      host/port defined)

    - give an error if an attempt is made to start a transfer that does
      NOT support reading from multiple ports in parallel and the number
      of net_ports currently defined > 1

-- where does the data end up?
After configuring jive5ab like this:

    net_port = [<addr1>@]port1[=<suffix1>] : [<addr2>@]port2[=<suffix2>] ;

and "record = on : <name> ;" then two threads are started to capture the
data from the network (note: the "record = nthread : <nReaderThreads>
;" setting is ignored).

If no suffixes are specified all data will be recorded in the same scan,
otherwise the traffic is recorded in "<name>_ds<suffix1>" and
"<name>_ds<suffix2>" respectively.

-- datastreams

The code supports reading "datastreams" enabled configurations from
multiple ports, provided no "=suffix" is set in the "net_port" command
(this is checked by the code); the recording suffixes will be taken from
the datastreams configuration instead. Thus it is possible to uniquely
filter on VDIF frame properties even if the streams are sent to
different destinations on the same recorder.

-- network packet statistics?
The e-vlbi packet statistics are now kept per stream (=net_port entry)
internally and the "evlbi?" query outputs the totals. There is scope to
extend the "evlbi?" query (or command) to output per-stream statistics
but no interface for that has been designed.

-- code reorganisation and templating
The src/threadfns.cc file had long since grown to a disproportional size
(> 6kLOC!) so there's always been a desire to move the individual thread
functions ("processing chain steps") to their own implementation files;
the "threadfns/" directory has also been around for a LOOOONG time, only
being populated by a few files so far.

The feature implementation prompted rewriting some of the processing
steps as templates anyway (so as to be able to transparently push blocks
or tagged blocks), so taking the code out of the .cc file and putting it
in a header file was necessary anyway, so why not relocate to a better
place whilst at it. The src/CMakeLists.txt was modified accordingly.

Some helper structs and typedefs had to be taken out of the
src/threadfns.cc file to be able to be included in different compilation
units, so, in order to avoid triggering a cyclic include problem, were
put in their own (very small) files or to more appropriate already
existing header file (such as moving the scoped_fd utility to
src/getsok.{h,cc}

Whilst at it fixed the include guards for the files in the
src/threadfns/ directory to be more consistent.