Fix warnings in both C++11 and non-C++11 builds
1) countedpointer.h

Newer compilers already complained about countedpointer.h reInterpret()
function using "placement new" pseudo-constructor construction we used
to make the cPtrBlock only increment the refcount (it was reinterpreting
the memory of a cPtrBlock to another type, adding a reference to derived
class so only the refcnt needs to be incremented).
But then you get the warning "Unitialized memory access" (which isn't
true but probably the compiler cannot guarantee either).

Fixed by using a different construction altogether.

2) src/stringutil.h, src/mk5command/group_def.cc

g++-13 complained about group_def's keypairextractor's operator= hiding
base class's operator=. And adding a "using ostringiterator::operator="
initially failed because in that base class the self-assignment operator
was deleted (C++11) or just declared-but-not-implemented (non-C++11),
both in the private section. Which made the compiler warn about the
operator= being private. Sigh.

Fixed by putting the "operator=" in the public section, using C++11
"=delete" to explicitly remove that operator; put the default c'tors
(which we also don't want) in the private section and use "=delete"
(C++11) or the declare-but-dont-define paradigm for non-C++11

So now the code base compiles fine again in both C++11 and non-C++11
mode on compilers from g++ 4.7.2 (Debian 4.7/amd64) to g++ 13 (Mac OSX 13.4.1/M1 Pro)