Special case if src/dst signature match
In order to help casting pointer-to-function-with-signature-A into
pointer-to-function-with-signature-B there was now a case where source and
destination signatures matched.  (In order to support ARM64 the target function
signature "void (*)(...)" didn't work; e.g. calling as "void(*)(arg1, arg2,
arg3, arg4)" it passed the function arguments arg1..arg4 only on x86; on ARM64
(Apple Silicon) the target function only received null pointers.

So the target calling convention for splitfunctions was set to "void (*)(void*,
uint, void* [x16])", but it meant that IF there was a source function with the
same signature (which happens to exist on x86/32bit platforms) then the casting
code got confused; the "template <A,B> ..." only worked correctly if A != B.
Now there is a special case for the converter ("do nothing") in case A == B