Local fix for matching chunks
Some station(s) are creating VDIF file names with an extra . in the name. This makes the chunk-regex match fail, leading to some issues with scan_set for example.

At the moment believe this patch to be benign but don't want to roll it out globally before having verified it doesn't break someone else's operations