JayB: ur datastream parsin' is kaputt (fixed)
Quoth user: "These ought to work Just Fine (TM) according to the spec but them don't no work"

    datastream = add : y1 : *@*/Y3.*
    !datastream = 4 : /Users/verkouter/src/jive5ab/src/mk6info.cc@874 assertion [match] fails The vdif criterion '*@*/Y3.*' does not match the specification;

    datastream = add : y1 : Y1.*
    !datastream = 4 : /Users/verkouter/src/jive5ab/src/mk6info.cc@874 assertion [match] fails The vdif criterion 'Y1.*' does not match the specification;

Problem reproducible, i.e. legit. Fixed by regex101.com help in debugging the proper expression to not fail to match.
The parsing code had to be modified a bit - the fields of the expression that needed to be decoded shifted a bit, and previous "host" and "port" were independent fields but now it's either "host@port" or just "port", so the logic had to cope with that.

Seems to parse a lot more and succesfully now.