Fix SIGSEGV
Several users (JaviG, JonQ, SimonC) wrote in that v3.1.0 might crash with a SIGSEGV after an
undeterminate amount of time.

With the help of SimonC this was traced to a memory race condition
triggered by "record=off".

All coredumps inspected (thanks @JaviG, @SimonC) crashed in the same
way, crashing in L6211.

    (gdb) list
    6210        if( fdreader->threadid!=0 ) {
    6211            int rv = ::pthread_kill(*fdreader->threadid, SIGUSR1);

    (gdb) print fdreader->threadid
    $2 = (pthread_t *) 0x0
    (gdb)

So it would seem that after testing for threadid!=0, the phtread_kill()
got fed a threadid of 0x0

Typical race condition: one thread modifying fdreader->threadid between
this thread's checking the value and using it.