First go fixing erroneous remvoval non-empty file
Users write in that net2file erroneously removes a file that it believes
to be empty.
The root cause is that the file descriptor gets closed before the guard
function checks it, then attempts to get the file size of a closed file,
which (erroneously?) returns 0 in stead of throwing an exception.

For now let's attempt by not adding an explicit close_filedescriptor and
see what happens