Need to keep std::atomic_init() happy about inputs
HarunaF and others wrote in:

'atomic_init(refcount_type*&, int)'
     std::atomic_init(refcountptr, 1);
                                    ^
In file included from .../src/block.h:31,
                 from .../src/block.cc:20:
/usr/include/c++/8/atomic:1062:5: note: candidate: 'template<class _ITp> void std::atomic_init(std::atomic<_ITp>*, _ITp)'
     atomic_init(atomic<_ITp>* __a, _ITp __i) noexcept
     ^~~~~~~~~~~
/usr/include/c++/8/atomic:1062:5: note:   template argument deduction/substitution failed:
.../src/block.cc:81:36: note:   deduced conflicting types for parameter '_ITp' ('unsigned int' and 'int')
     std::atomic_init(refcountptr, 1);

I.e. std::atomic_init() needs to be presented with an initializer of the
correct type.

In C++11 we know "refcount_type" is std::atomic<...> so made the inits
do a static typecast following whatever the underlying storage type of
the refcount_type actually is.